+++ gui/dc/image/views.py
@@ -159,15 +159,31 @@ def imagestore_list(request, repo=None):
     context['form_admin'] = AdminImageForm(request, None, prefix='adm', initial={'owner': user.username,
                                                                                  'access': Image.PRIVATE,
                                                                                  'dc_bound': True})
+    qs_image_filter = request.GET.copy()
+    qs_image_filter.pop('created_since', None)
+    qs_image_filter.pop('last', None)
+    context['qs_image_filter'] = qs_image_filter.urlencode()
+    context['default_limit'] = default_limit = 30
 
     try:
-        created_since_days = int(request.GET.get('created_since', '30'))
+        created_since_days = int(request.GET.get('created_since', 0))
     except (ValueError, TypeError):
-        created_since_days = 30
+        created_since_days = None
+
+    if created_since_days:
+        limit = None
+    else:
+        created_since_days = None
+
+        try:
+            limit = int(request.GET.get('last', default_limit))
+        except (ValueError, TypeError):
+            limit = default_limit
 
     repositories = ImageStore.get_repositories()
     context['imagestores'] = imagestores = ImageStore.all(repositories)
     context['created_since'] = created_since_days
+    context['limit'] = limit
 
     if repositories:
         if repo and repo in repositories:
@@ -178,9 +194,9 @@ def imagestore_list(request, repo=None):
         if created_since_days:
             created_since = make_aware(datetime.now() - timedelta(days=created_since_days))
         else:
-            created_since = ''
+            created_since = None
 
-        context['images'] = imagestore.images_filter(created_since=created_since)
+        context['images'] = imagestore.images_filter(created_since=created_since, limit=limit)
     else:
         context['imagestore'] = None
         context['images'] = []